feat: 修改目录结构，新增service、layouts、types。修改组件为ts类型，添加错误代码注释。
