--- src/components/Footer.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import { Layout } from 'antd';
-
-const { Footer } = Layout;
-
-
-export default function Footers() {
-  return (
-    <Footer style={{ textAlign: 'center' }}>Ant Design ©2018 Created by Ant UED</Footer>
-  )
-}
