+++ src/index.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 import { HashRouter } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(
@@ -15,5 +14,3 @@ root.render(
     </HashRouter>
   </React.StrictMode>
 );
-
-reportWebVitals();
