+++ src/layouts/Footer.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Layout } from 'antd';
+
+const { Footer } = Layout;
+
+const Footers: React.FC = () => {
+  return (
+    <Footer style={{ textAlign: 'center' }}>Ant Design ©2018 Created by Ant UED</Footer>
+  )
+}
+export default Footers;
