+++ src/pages/Login/Login.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Login: React.FC = () => {
+  return (
+    <div>
+      login
+    </div>
+  )
+}
+
+export default Login;
