+++ src/pages/SandBox/BlogManage/BlogAdd.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const BlogAdd: React.FC = () => {
+  return (
+    <div>
+      BlogAdd
+    </div>
+  )
+}
+
+export default BlogAdd;
