+++ src/pages/SandBox/BlogManage/BlogDraft.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const BlogDraft: React.FC = () => {
+  return (
+    <div>
+      BlogDraft
+    </div>
+  )
+}
+
+export default BlogDraft;
