+++ src/pages/SandBox/BlogManage/BlogPublished.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const BlogPublished: React.FC = () => {
+  return (
+    <div>
+      BlogPublished
+    </div>
+  )
+}
+
+export default BlogPublished;
