+++ src/pages/SandBox/Category/Category.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Category: React.FC = () => {
+  return (
+    <div>
+      Category
+    </div>
+  )
+}
+
+export default Category;
