+++ src/pages/SandBox/Home/Home.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Home: React.FC = () => {
+  return (
+    <div>
+      home
+    </div>
+  )
+}
+
+export default Home
