+++ src/types/Login/Login.tsx
@@ -1,0 +1,1 @@
+export default 1;
