+++ src/types/SandBox/SandBox.tsx
@@ -1,0 +1,1 @@
+export default 1;
