--- src/views/Login/Login.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Login() {
-  return (
-    <div>
-      login
-    </div>
-  )
-}
