--- src/views/SandBox/Category/Category.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Category() {
-  return (
-    <div>
-      Category
-    </div>
-  )
-}
