--- src/views/SandBox/Home/Home.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Home() {
-  return (
-    <div>
-      home
-    </div>
-  )
-}
