--- src/views/SandBox/blog_manage/BlogAdd.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function BlogAdd() {
-  return (
-    <div>
-      BlogAdd
-    </div>
-  )
-}
