--- src/views/SandBox/blog_manage/BlogDraft.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function BlogDraft() {
-  return (
-    <div>
-      BlogDraft
-    </div>
-  )
-}
