--- src/views/SandBox/blog_manage/BlogPublished.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function BlogPublished() {
-  return (
-    <div>
-      BlogPublished
-    </div>
-  )
-}
