--- src/views/SandBox/blog_manage/BlogUnPublish.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function BlogUnPublish() {
-  return (
-    <div>
-      BlogUnPublish
-    </div>
-  )
-}
